+++ core/src/routes/(modules)/actions/div.svelte
@@ -1,12 +1,14 @@
 <script>
     import "../../../app.css";
+	import Iconbutton from "../../uicomponents/iconbutton.svelte";
+    import { globalThemeStore } from "../../globals/globalstores.js";
 
     /**
      * This is a default variable, which holds position of "Actions" panel according to top-right corner on X axis.
      * 
      * Do not change variable name, since this is a standart definition for all modules.
      */
-    export const marginX = "-6px";
+    export const marginX = "-20px";
 
     /**
      * This is a default variable, which holds position of "Actions" panel according to top-right corner on Y axis.
@@ -25,10 +27,34 @@
 
 </script>
 
-<div class="flex flex-row w-16 gap-1 divide-x divide-black/40 place-content-center items-center content-center">
-    <div><button class="bg-transparent border-none w-6 h-6 p-0 m-0 text-black" on:click='{editButtonPress}'><i class="bi bi-plus w-5 h-5 text-black"></i></button></div>
-    <div><button class="bg-transparent border-none w-6 h-6 p-0 m-0 text-black" on:click='{editButtonPress}'><i class="bi bi-three-dots-vertical w-5 h-5 text-black"></i></button></div>
-    
+<div class="flex flex-row w-28 rounded-md gap-1 divide-x divide-black/40 place-content-center items-center content-center inlinePanel" style='
+--fixedPanelBackgroundColor:{$globalThemeStore.panel.backgroundColor};
+--fixedPanelForegroundColor:{$globalThemeStore.panel.foregroundColor};
+--fixedPanelTitleColor:{$globalThemeStore.panel.titleColor};
+--fixedPanelTabsDivider: {$globalThemeStore.panel.tabsDivider};
+--backgroundColor: {$globalThemeStore.widgetIcon.backgroundColor};
+--foregroundColor: {$globalThemeStore.widgetIcon.foregroundColor};
+--borderColor: {$globalThemeStore.widgetIcon.borderColor};
+--iconColor: {$globalThemeStore.widgetIcon.iconColor};
+--textColor: {$globalThemeStore.widgetIcon.textColor};
+' >
+
+    <div class="flex flex-row gap-1">
+        <Iconbutton active={false} on:click={editButtonPress} noBackground><span slot="icon"><i class="bi bi-pen"></i></span></Iconbutton>
+        <Iconbutton active={false} on:click={editButtonPress} noBackground><span slot="icon"><i class="bi bi-files"></i></span></Iconbutton>
+    </div>
+
+    <div>
+        <Iconbutton active={false} on:click={editButtonPress} noBackground><span slot="icon"><i class="bi bi-x-circle"></i></span></Iconbutton>
+    </div>
     
 </div>
 
+<style>
+
+    :global(.inlinePanel){
+        background-color: var(--fixedPanelBackgroundColor);
+        color: var(--fixedPanelForegroundColor);
+    }
+
+</style>